+++ CHANGELOG.md
@@ -1,3 +1,7 @@
+0.5.8.3
+=======
+* GHC 9.4.1 compatibility
+
 0.5.8.1
 =======
 * GHC 8.8.1 compatibility